Impl. message transmitters
