+++ src/main/java/org/radarbase/appserver/entity/Notification.java
@@ -46,7 +46,7 @@ import org.springframework.lang.Nullable;
  *
  * @author yatharthranjan
  * @see Scheduled
- * @see org.radarbase.appserver.service.scheduler.NotificationSchedulerService
+ * @see org.radarbase.appserver.service.scheduler.MessageSchedulerService
  */
 @Table(
         name = "notifications",
