+++ src/main/java/org/radarbase/appserver/service/FcmNotificationService.java
@@ -42,7 +42,7 @@ import org.radarbase.appserver.exception.NotificationAlreadyExistsException;
 import org.radarbase.appserver.repository.NotificationRepository;
 import org.radarbase.appserver.repository.ProjectRepository;
 import org.radarbase.appserver.repository.UserRepository;
-import org.radarbase.appserver.service.scheduler.NotificationSchedulerService;
+import org.radarbase.appserver.service.scheduler.MessageSchedulerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -66,7 +66,7 @@ public class FcmNotificationService implements NotificationService {
     private final transient NotificationRepository notificationRepository;
     private final transient UserRepository userRepository;
     private final transient ProjectRepository projectRepository;
-    private final transient NotificationSchedulerService schedulerService;
+    private final transient MessageSchedulerService schedulerService;
     private final transient NotificationConverter notificationConverter;
     private final transient ApplicationEventPublisher notificationStateEventPublisher;
 
@@ -75,7 +75,7 @@ public class FcmNotificationService implements NotificationService {
             NotificationRepository notificationRepository,
             UserRepository userRepository,
             ProjectRepository projectRepository,
-            NotificationSchedulerService schedulerService,
+            MessageSchedulerService schedulerService,
             NotificationConverter notificationConverter,
             ApplicationEventPublisher eventPublisher) {
         this.notificationRepository = notificationRepository;
