+++ src/main/java/org/radarbase/appserver/service/transmitter/DataMessageTransmitter.java
@@ -1,0 +1,9 @@
+package org.radarbase.appserver.service.transmitter;
+
+import org.radarbase.appserver.entity.DataMessage;
+import org.radarbase.appserver.exception.MessageTransmitException;
+
+public interface DataMessageTransmitter {
+    // Can Sneakily throw MessageTransmitException
+    void send(DataMessage dataMessage);
+}
