+++ src/main/resources/application-prod.properties
@@ -40,6 +40,17 @@ spring.quartz.properties.auto-startup=true
 #spring.quartz.properties.org.quartz.jobStore.useProperties = true
 spring.quartz.properties.org.quartz.jobStore.driverDelegateClass=org.quartz.impl.jdbcjobstore.PostgreSQLDelegate
 
+# EMAIL SETTINGS
+# needed when notifications via email are enabled
+notification.email.enabled=true
+notification.email.from=radar@radar.thehyve.nl
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.username=<login user to smtp server>
+spring.mail.password=<login password to smtp server>
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
+
 # LOGGING
 logging.level.root=INFO
 logging.level.org.springframework.web=INFO
